Moved validation support code out of case service tests into its own file
