+++ hrm-service/.eslintrc.js
@@ -25,5 +25,5 @@ module.exports = {
     ecmaVersion: 2018, // Allows for the parsing of modern ECMAScript features
     sourceType: 'module', // Allows for the use of imports
   },
-  ignorePatterns: ['dist/**'],
+  ignorePatterns: ['dist/**', 'coverage/**'],
 };
